Flow root libraries intent through child libs
One suggestion from ChrisMc was that we should have a way for sub-libraries to use the root libraries intents. With this change the root libraries recognizers will be run and then the top intent will be passed down through to sub libraries.  Any library that doesn't setup recognizers will use the root libraries top intent which should be safe.

There was some logic added to avoid running the root libs recognizers twice. I also updated the various Library.findRoutes() methods to take a read-only recognizer context instead of a session which is general goodness.